LayoutTests/accessibility/accessibility-object-model.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
<script src="../resources/accessibility-helper.js"></script>
</head>
<body>

<button id="button">Click Me</button>
<div id="container"><button id="button2">Button2</button></div>
<p id="paragraph"></p>

<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests getting and setting Accessibility Object Model properties.");
    if (window.accessibilityController) {
        var button = document.getElementById("button");
        var axButton = accessibilityController.accessibleElementById("button");
        var aomRemovedButton;
        var paragraph = document.getElementById("paragraph");
        var axParagraph;
        
        shouldBeFalse("button.accessibleNode == null");
        
        testPropertiesDefault();
        testNoReflection();
        testSettingProperties();
        testBecomeAccessible();
        testInvalidRole();
        testElementAlive();
    }
    
    function testPropertiesDefault() {
        debug("\nSupported properties on an AccessibleNode are all null by default");
        shouldBeNull("button.accessibleNode.role");
        shouldBeNull("button.accessibleNode.label");
        // Invalid property value should be undefined
        shouldBe("button.accessibleNode.foo", "undefined");
    }
    
    function testNoReflection() {
        debug("\nARIA attributes should not be reflected into AOM properties.");
        button.setAttribute("role", "checkbox");
        button.setAttribute("aria-label", "label");
        shouldBe("axButton.role", "'AXRole: AXCheckBox'");
        shouldBe("axButton.description", "'AXDescription: label'");
        
        // AOM properties should be null even if we have set ARIA attributes.
        shouldBeNull("button.accessibleNode.role");
        shouldBeNull("button.accessibleNode.label");
    }
    
    function testSettingProperties() {
        debug("\nTest setting AOM properties. And make sure AOM takes precedence.");
        
        // Set the ARIA attributes on the element first.
        button.setAttribute("role", "checkbox");
        button.setAttribute("aria-label", "label");
        
        // Then set the corresponding AOM properties to some different values.
        button.accessibleNode.role = "slider";
        shouldBe("button.accessibleNode.role", "'slider'");
        button.accessibleNode.label = "AOM Label";
        shouldBe("button.accessibleNode.label", "'AOM Label'");
        
        // The AOM property values should override ARIA attributes.
        shouldBe("axButton.role", "'AXRole: AXSlider'");
        shouldBe("axButton.description", "'AXDescription: AOM Label'");
    }
    
    function testBecomeAccessible() {
        debug("\nSetting some of the AOM properties should be able to make an element accessible.");
        axParagraph = accessibilityController.accessibleElementById("paragraph");
        shouldBeTrue("axParagraph == null || axParagraph == undefined");
        
        // The element should be accessible if it has a label.
        paragraph.accessibleNode.label = "test label";
        axParagraph = accessibilityController.accessibleElementById("paragraph");
        shouldBeFalse("axParagraph.isIgnored");
    }
    
    function testInvalidRole() {
        debug("\nAn invalid role should be ignored.");
        
        // Clear the ARIA attribute and AOM property value.
        button.removeAttribute("role");
        button.accessibleNode.role = null;
        shouldBe("button.accessibleNode.role", "null");
        shouldBe("axButton.role", "'AXRole: AXButton'");
        
        // Accessibility should use the semantic role if an invalid role is provided.
        button.accessibleNode.role = "badrole";
        shouldBe("button.accessibleNode.role", "'badrole'");
        shouldBe("axButton.role", "'AXRole: AXButton'");
    }
    
    function testElementAlive() {
        debug("\nAn AccessibleNode keeps its element alive.");
        // Get the button to be removed and access its accessibleNode.
        (function() { 
            var button2 = document.getElementById("button2");
            aomRemovedButton = button2.accessibleNode;
            aomRemovedButton.role = "checkbox";
        })();
        shouldBe("aomRemovedButton.role", "'checkbox'");
        
        // Remove the button make sure we are still able to access the accessibleNode.
        (function() {
            var button2 = document.getElementById("button2");
            button2.parentElement.removeChild(button2);
        })();
        gc();
        shouldBe("aomRemovedButton.role", "'checkbox'");
    }
    
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
