LayoutTests/accessibility/accessibility-object-model-expected.txt

Click Me
This tests getting and setting Accessibility Object Model properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS button.accessibleNode == null is false

Supported properties on an AccessibleNode are all null by default
PASS button.accessibleNode.role is null
PASS button.accessibleNode.label is null
PASS button.accessibleNode.foo is undefined

ARIA attributes should not be reflected into AOM properties.
PASS axButton.role is 'AXRole: AXCheckBox'
PASS axButton.description is 'AXDescription: label'
PASS button.accessibleNode.role is null
PASS button.accessibleNode.label is null

Test setting AOM properties. And make sure AOM takes precedence.
PASS button.accessibleNode.role is 'slider'
PASS button.accessibleNode.label is 'AOM Label'
PASS axButton.role is 'AXRole: AXSlider'
PASS axButton.description is 'AXDescription: AOM Label'

Setting some of the AOM properties should be able to make an element accessible.
PASS axParagraph == null || axParagraph == undefined is true
PASS axParagraph.isIgnored is false

An invalid role should be ignored.
PASS button.accessibleNode.role is null
PASS axButton.role is 'AXRole: AXButton'
PASS button.accessibleNode.role is 'badrole'
PASS axButton.role is 'AXRole: AXButton'

An AccessibleNode keeps its element alive.
PASS aomRemovedButton.role is 'checkbox'
PASS aomRemovedButton.role is 'checkbox'
PASS successfullyParsed is true

TEST COMPLETE
